Titanic Syndrome
A small-scale C# test program. The test is used to determine the users' level of risk of suffering from the Titanic Syndrome, based on a scale from 0-75. The program is designed so that the users who are a part of an organization can take the test based on their working experience at the business, or they can take it for themselves. All of the users' data is stored to an excel file on completion of the test, and the administrator of the program is able to login with their particulars to view the statistics on all of the people who have taken the test.